// src/lib/http/apis/users.d.ts
export type ProfileType = {
  id: number | string
  first_name: string
  last_name: string
  middle_name: string
  display_name: string
  birthday: string
  email: string
  phone_number: string
  address: string
  sex: string
  national_id: string
  nation: string
  owner: number
  creator: number
}